# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Học Tiếng Anh",
        "description": "Tài liệu API cho ứng dụng học tiếng Anh.",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "fb66b153966d9fac2e999136ef0ade67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không hợp lệ"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký người dùng mới",
                "operationId": "fb835c97347e657e2a627790c4c5817a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Người dùng đã được tạo thành công"
                    },
                    "422": {
                        "description": "Lỗi xác thực"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng xuất người dùng",
                "operationId": "307ed0512855922b2280792376bb2095",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Thay đổi mật khẩu người dùng",
                "operationId": "bd4dd230facd62e300c1bed2cc7f512f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thay đổi mật khẩu thành công"
                    },
                    "400": {
                        "description": "Lỗi xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/get-all": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Danh sách danh mục (đơn giản)",
                "operationId": "50e6bb45bd54574e535faa878d0a6efe",
                "responses": {
                    "200": {
                        "description": "Danh sách danh mục đơn giản"
                    }
                }
            }
        },
        "/api/v1/categories": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Tạo danh mục mới",
                "operationId": "3d09a8f4aea96aed370bc0120547a197",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Sách tiếng Anh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo danh mục thành công"
                    },
                    "400": {
                        "description": "Tạo danh mục thất bại"
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Cập nhật danh mục",
                "operationId": "8edd529bf8c5d2c024f479805defe8b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Sách luyện nghe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            }
        },
        "/api/v1/auth/profile": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "summary": "Lấy thông tin người dùng",
                "operationId": "796834ad6f77e404063cd009279feb5d",
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Category",
            "description": "Category"
        },
        {
            "name": "UserProfile",
            "description": "UserProfile"
        }
    ]
}